Introduce ArraySerializer class
